requirements.txt
openai>=1.0.0,<2.0.0
requests>=2.31.0,<3.0.0
tzdata>=2024.1
python-dotenv>=1.0.0,<2.0.0
